+++ demisto_sdk/commands/secrets/secrets.py
@@ -107,8 +107,8 @@ class SecretsValidator(object):
                                             ' remove the files asap and report it.\n'
 
                 secrets_found_string += 'For more information about whitelisting visit: ' \
-                                        'https://github.com/demisto/demisto-sdk/tree/master/demisto_sdk/' \
-                                        'commands/secrets'
+                                        'https://pull-request-696--demisto-content-docs.netlify.app/docs/concepts/demisto-sdk' \
+                                        '#secrets'
                 print_error(secrets_found_string)
         return secret_to_location_mapping